move this out of the common mod initializer
