+++ src/main/java/com/wildfire/main/WildfireGender.java
@@ -18,49 +18,26 @@
 
 package com.wildfire.main;
 
-import java.time.Duration;
-import java.util.*;
-
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
-import com.mojang.brigadier.arguments.IntegerArgumentType;
-import com.mojang.brigadier.arguments.StringArgumentType;
-import com.mojang.brigadier.context.CommandContext;
 import com.mojang.logging.LogUtils;
 import com.wildfire.main.config.ClientConfig;
-import com.wildfire.main.config.enums.SyncVerbosity;
-import com.wildfire.main.entitydata.EntityConfig;
 import com.wildfire.main.entitydata.PlayerConfig;
 import com.wildfire.main.networking.WildfireSync;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.client.command.v2.ClientCommandManager;
-import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
-import net.fabricmc.fabric.api.client.command.v2.FabricClientCommandSource;
-import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.projectile.ProjectileUtil;
-import net.minecraft.server.MinecraftServer;
-import net.minecraft.server.command.CommandManager;
-import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.text.Text;
-import net.minecraft.util.hit.EntityHitResult;
-import net.minecraft.util.math.Vec3d;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 
-import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.argument;
+import java.time.Duration;
+import java.util.List;
+import java.util.UUID;
 
 public class WildfireGender implements ModInitializer {
 	public static final String MODID = "wildfire_gender";
 	public static final Logger LOGGER = LogUtils.getLogger();
 	public static final LoadingCache<UUID, PlayerConfig> CACHE;
-
-	//TODO: Move to client cfg
-	public static boolean DEBUG_MODE = true;
 
 	static {
 		var builder = CacheBuilder.newBuilder();
@@ -106,131 +83,7 @@ public class WildfireGender implements ModInitializer {
 		WildfireSync.register();
 		WildfireEventHandler.registerCommonEvents();
 		ClientConfig.INSTANCE.load();
-
-
-		ClientCommandRegistrationCallback.EVENT.register((dispatcher, registryAccess) -> {
-			dispatcher.register(
-					ClientCommandManager.literal("fgm")
-					.then(ClientCommandManager.literal("invalidatecache")
-							.executes(WildfireGender::invalidateCache)
-					)
-					.then(ClientCommandManager.literal("lookentity")
-							.executes(WildfireGender::getEntityLookingAt)
-					)
-					.then(ClientCommandManager.literal("cache")
-							.executes(WildfireGender::getUsers)
-					)
-					.then(ClientCommandManager.literal("debug")
-							.executes(WildfireGender::debugCommand)
-					)
-					.then(ClientCommandManager.literal("verbosity")
-						.then(argument("level", StringArgumentType.string())
-								.suggests((ctx, builder) -> {
-									for (SyncVerbosity level : SyncVerbosity.values()) {
-										builder.suggest(level.name());
-									}
-									return builder.buildFuture();
-								})
-								.executes(WildfireGender::setLogLevel)
-						)
-					)
-			);
-		});
 	}
-
-	private static int getEntityLookingAt(CommandContext<FabricClientCommandSource> ctx) {
-		MinecraftClient client = MinecraftClient.getInstance();
-		if (client.player == null || client.world == null) return 0;
-
-		double reachDistance = 10.0D; // how far to check
-		Vec3d eyePos = client.player.getCameraPosVec(1.0F);
-		Vec3d lookVec = client.player.getRotationVec(1.0F);
-		Vec3d reachVec = eyePos.add(lookVec.multiply(reachDistance));
-
-		EntityHitResult entityHitResult = ProjectileUtil.raycast(
-				client.player,
-				eyePos,
-				reachVec,
-				client.player.getBoundingBox().stretch(lookVec.multiply(reachDistance)).expand(1.0D),
-				entity -> !entity.isSpectator() && entity.isAlive(),
-				reachDistance * reachDistance
-		);
-
-		if (entityHitResult != null && entityHitResult.getEntity() != null) {
-			Entity target = entityHitResult.getEntity();
-			WildfireHelper.logCommand(ctx, "Looking at: " + target.getName().getString());
-			WildfireHelper.logCommand(ctx, "UUID: " + target.getUuidAsString());
-			WildfireHelper.logCommand(ctx, "Type: " + target.getType());
-			WildfireHelper.logCommand(ctx, "Class: " + target.getClass());
-			WildfireHelper.logCommand(ctx, "Renderer: " + MinecraftClient.getInstance().getEntityRenderDispatcher().getRenderer(target));
-
-		} else {
-			WildfireHelper.logCommand(ctx, "No entity in sight.");
-		}
-		return 1;
-	}
-
-	//COMMANDS BELOW
-	public static int setLogLevel(CommandContext<FabricClientCommandSource> ctx) {
-		String level = StringArgumentType.getString(ctx, "level");
-
-		ClientConfig.INSTANCE.set(ClientConfig.SYNC_VERBOSITY, SyncVerbosity.valueOf(level));
-		ClientConfig.INSTANCE.save();
-
-		WildfireHelper.logCommand(ctx, "Log level set to: " + level);
-		return 1;
-	}
-
-	private static int getUsers(CommandContext<FabricClientCommandSource> ctx) {
-		if (MinecraftClient.getInstance().world != null) {
-			ctx.getSource().sendFeedback(Text.literal("Players Cached (" + WildfireGender.CACHE.size() + "):"));
-
-			for (Map.Entry<UUID, PlayerConfig> entry : WildfireGender.CACHE.asMap().entrySet()) {
-				PlayerConfig plrConfig = entry.getValue();
-				if (plrConfig != null) {
-					PlayerEntity plr = MinecraftClient.getInstance().world.getPlayerByUuid(plrConfig.uuid);
-					if (plr != null) {
-						ctx.getSource().sendFeedback(
-								Text.empty().append(plr.getDisplayName()).append(" - ").append(plrConfig.getGender().getDisplayName())
-						);
-					}
-				}
-			}
-
-			ctx.getSource().sendFeedback(Text.literal("Entities Cached (" + EntityConfig.CACHE.size() + "):"));
-
-			for (Map.Entry<UUID, EntityConfig> entry : EntityConfig.CACHE.asMap().entrySet()) {
-				EntityConfig entityConfig = entry.getValue();
-				if (entityConfig != null) {
-					Entity entity = MinecraftClient.getInstance().world.getEntity(entityConfig.uuid);
-					if (entity != null) {
-						ctx.getSource().sendFeedback(
-								Text.empty().append(entity.getDisplayName()).append(" - ").append(entityConfig.getGender().getDisplayName())
-						);
-					}
-				}
-			}
-		}
-		return 1;
-	}
-
-
-	private static int invalidateCache(CommandContext<FabricClientCommandSource> ctx) {
-		CACHE.invalidateAll();
-		EntityConfig.CACHE.invalidateAll();
-
-		WildfireHelper.logCommand(ctx, "Cache has been invalidated!");
-		return 1;
-	}
-
-	private static int debugCommand(CommandContext<FabricClientCommandSource> ctx) {
-		ClientConfig.INSTANCE.set(ClientConfig.DEBUG_MODE, !ClientConfig.INSTANCE.get(ClientConfig.DEBUG_MODE));
-		WildfireHelper.logCommand(ctx, "Debug Mode: " + (ClientConfig.INSTANCE.get(ClientConfig.DEBUG_MODE)?"Enabled":"Disabled"));
-		ClientConfig.INSTANCE.save();
-		return 1;
-	}
-
-	//END COMMANDS BELOW
 
 	public static @Nullable PlayerConfig getPlayerById(UUID id) {
 		return CACHE.getIfPresent(id);
